requirements.txt
pynput 
python-dotenv
requests
pynacl